Update quickstart for latest gerrit
Gerrit has removed the ability for the admin user to directly
push to All-Projects by default, which hinders our ability to
bootstrap the Gerrit configuration in the quickstart.  Update the
setup script to create a change instead and merge it.

This will have the unfortunate effect of leaving an actual change
in the system that users may end up seeing during the tutorial,
which may be distracting, but should not affect functionality.

Change-Id: I0316d1c2ab4f2dceccf18adc2cb2fb283e3b2226