Reduce ZK lock contention in executor
Now that the job request queue uses the tree cache, store
information on whether or not each request is locked on our
cached objects.  Use that to avoid attempting to lock a request
that our cache thinks is already locked.

This should reduce executor contention somewhat since as soon as
one executor locks a request, other executors can avoid
contributing to the thundering herd of lock attempts.

Change-Id: I0317eff129622858285359d3a857044a8317ccf5