Use rendezvous hashing for executor distribution
This replaces our current gearman-inspired delay-based load
distribution with rendezvous hashing.

Each build request is hashed across the set of executors in that
build's zone that are accepting jobs.  The lowest hash value is
the one that will be expected to run the job.  No other executors
will attempt to run it.

The executors may be slightly out of sync with their knowledge of
which ones are accepting jobs.  In the case that multiple executors
think they should run the job, locking will resolve the conflict.
In the case that no executors think they should run the job, we
subscribe to events from the component registry, and any time some
executor changes state, we will re-scan the list of requests,
recompute hashes, and run the job if appropriate.

The current system is fairly good at quickly evening out the load
distribution when new executors come online.  It is not clear if
the same will be true for this method.  Since the current system
delays more as more jobs are running, unloaded executors accept
jobs quickly.  This system will evenly distribute them across
all accepting executors, meaning that a new executor may accept
a smaller number of new jobs compared to the current system.
However, the accepting jobs flag is still a contributing factor,
and if existing executors stop accepting jobs (perhaps because
their starting-jobs governor takes them offline) more jobs will
be directed to the new executor(s).

Change-Id: I8b12fe8fa0b3f4e8ca5d21d2ee05be4269c677ef