Use Rendezvous hashing for merger jobs
Building on the change to use rendezvous hashing to distribute
build jobs to executors, do the same for merge jobs.

This should reduce ZK traffic and lock contention, especially in
systems with more mergers, though it may distribute jobs less
efficiently; if there is a small queue of jobs, we may end up with
idle mergers while busy mergers have a queue.

Change-Id: I2e4a50b4fa894374de636f55e8d37c0f7c753151