Add ability to mark individual job attributes as final
This adds a new job configuration attribute which allows users
to mark individual job attributes as final.  For example, a user
may want to mark required-projects as final in order to ensure
that updates to add new projects are always made to a parent job.

This is similar to override control, but is implemented as a
dictionary rather than yaml tags because there is limited space
to accomodate yaml tags, and this is expected to be used less
frequently and more deliberately than override control.

The dictionary is forward-compatible with other modifiers, such
as a potential expansion to mark individual attributes as
abstract or protected.

Change-Id: Ia71fd286cf84b7bf449f6442f8734abc41734050