Parallelize workspace repo restoration
When preparing workspace repos, we restore the repo state using
very efficient python code in-process.  Despite the efficiency,
it is still python code in the main process which means there
is an opportunity for further efficiency by offloading it to
a subprocess.

This change moves that work into the processpool that the executor
maintains for offloading git repo work.  It also performs
restoration for all of the involved repos in parallel.

To accomplish this, executors do the following when starting
a build:
* Clone each of the repos (in series)
* Restore the workspace of all repos (in parallel)
* Merge changes (in series; only affects some repos)

There may be opportunity for further parallelization, but we do
want to balance work done for an individual job against the rest
of the executors workload.  We should evaluate this change before
committing to further parallelization.

Change-Id: I2727b00d540ed0d6d650342f427503ddcf5cd533