Push and approve Gerrit ACL update in one command
Since Gerrit 3.11 we cannot directly push ACL updates as the admin
Gerrit user. Instead we must push the changes for review and then code
review them. There is a shortcut where you can push the code and provide
a code review in the same command. Use this shortcut when setting Gerrit
up for the Zuul quickstart tutorial.

Change-Id: Icc417a4b6960aacf0625954bb142041c3460aa2e