Add provider list web API endpoint
This only returns a list of provider names right now.  We may
extend it in the future, but the main purpose at the moment is
to have a way to indicate to the web app that there are providers
in a given tenant and therefore it should show additional tabs
for provider-related pages.

This could have been accomplished with the tenant info endpoint,
but this is potentially more useful in the future and does not
need to be immeditaely deprecated.

Being the first NIZ related endpoint, it also sets up the
infrastructure in zuul-web for NIZ.

Change-Id: Ib3006c64a2e93061be6b29b62ec4f4ee0acb89f4