WIP: Allow pinning pipelines on status page
To give users more control over the order in which pipelines are
displayed on the pipeline overview, allow to mark them as "favourites",
so that they are displayed first and regardless of whether they are
empty or not.

This allows for users to personalize their view on the status page for
the pipelines they are most interested in. This is distinct from the
existing filtering feature. While filtering for pipelines also gives a
personalized view on the status page, this feature does not filter or
hide anything, but allows to sort pipelines based on personal
preference.

This change serves two goals:

1. In large systems/tenants with many pipelines, where some pipelines
   are used only by a subset of projects, it gives users a possibilty to
   more easily focus on specific pipelines without losing the view on
   others.
2. Reduce pipelines jumping around when others appear and disappear
   based on them being empty or not.

What is still undecided and left open in this change is where to
persistently store the pipeline pinning state (thus WIP). There are at
least two options which would need to be discussed and agreed upon
(other ideas are welcome).

1. Store and load the state for redux from the browsers localStorage

   This has a downside for users that clear their local storage when
   exiting their browser, rendering this feature rather useless with
   pins being gone after each session or also when switching to another
   browser. For other users, this might be the more comfortable
   solution, as state persistence is automatic.
   We already store other application state in the localStorage, e.g.
   the 'show all pipelines' toggle. However, that's a much simpler state
   for a user to restore if it would be gone.

2. Encode the state in the URL as query params like we do for filters

   This has the downside of more complex URLs (might be negligible) and
   users need to explicitly bookmark the URL with its state and update
   bookmarks after changes for persistence. While this solution requires
   far more user involvement, it would, on the other hand, allow to
   transfer the state across sessions and browsers.

Change-Id: Id58877de7040af642c89d162ae85d6273adfa081