Update patch set 1
Patch Set 1: Code-Review-1

Hi, thanks for coming up with this.  I like the mechanics of it and agree that it's intuitive.

I am concerned with the persistence.  We decided to keep the amount of persistent data used by the web app very small.  That's because many of our developers are quite privacy focused and do clear browser sessions as you described.  We also have developers that switch between many different workstations (including phones), so sharing between those browsers is important too.

We really don't want to store anything about the users on the servers.  It's just not necessary for what Zuul does.  But when we start trying to have user-customized dashboards, that's really the only way for things to behave as all users expect.

I'll note that even gerrit, which *does* store user data, has historically not had user-customized persistent dashboards and instead relied on users creating dashboards with search filters and then bookmarking the URLs.  So this workflow should not be foreign to Zuul users who use Gerrit, at least.

The effects of this change can *almost* be achieved today with the search filters.  Users can add a searhc filter for the pipelines they want; the only missing piece is "also include pipelines not in the filter but that do have changes in them".  It seems like we could probably achieve that with some kind of toggle.  We're getting too many toggles, so maybe we need to rethink that, but, structurally, it's sound.

There may be an alternative UI for this that is more like the current search filters, and that may make more sense for persisting the state in the URL.  Or we could change the UI to pinning as implemented in this change and persist the state in the URL.

Either way, if we want to have a solution like this, then we should consider persisting it in the URL because it does achieve all of the most important goals for all classes of users: the ability to create, share, and return to customized status pages.

To summarize the above: persistence mechanism and the UI for customizing pipeline display are orthogonal; we currently have chosen "filtering" for the UI, and "URL" for persistenc.

--------------

Some other brainstorm-level thoughts on persistence:

* We could allow for exporting/importing preferences by downloading files, or scanning qr codes, or something.  That might let us store state in local storage and still let users share between different workstations.

* We could leave this change as-is with no persistence and then everyone gets the same user experience.  I'm sure we will get comments about that from people who want to set up complicated filters and then open a new tab or exit their browser.

* Privacy-focused users can, of course, configure their browsers with exceptions for specific sites.  That may be a reasonable compromise, but does not address the issue of using multiple browsers.  During the recent status page work, we heard feedback from people that use 5 browsers regularly.  I sympathize; I use several as well.

---------------

Final thoughts:

I'd like to know more about the use case driving this.  Point #1 in the commit message is a good description of how a user wants to achieve something by customizing the set of pipelines displayed.  But I think it would be useful to know more about what it is they want to achieve.  The user described does not sound like the "operator who wants to find systemic problems" or the "developer who wants to track the progress of their change".  Perhaps some kind of tenant manager?  What are they looking for and why don't either of the current solutions (show an explicit set of pipelines, or show all the pipelines) work for them?

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, 4ddd7ae8cd3d6251ae9ea30f2c9e20329776ca2a
Attention: {"person_ident":"Gerrit User 31083 \u003c31083@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}