Incorporate tenants stats in Toolbar section
Currently, the queue length and the last reconfiguration timestamp are
always shown on top of the status page. Although this information is
quite important, it's not relevant most of the time and for most users.

So, instead of allocating a lot of space at the top of the page we could
utilize this space to show more of the "main" content and incorporate
the stats in the toolbar.

To show these different stats in a single place, they are visualized via
a dropdown. The dropdown can be opened by clicking on a stats button next
to the reload button in the top right corner of the page.

The design idea behind this change is that the toolbar is now split up
in two parts: A filter section on the left side and a button group on
the right side. The filter section controls which elements are shown on
the status page whereby the button group provides additional
functionality like refreshing the page or showing tenant-specific stats.

This makes the top of the status page much cleaner and more compact as
the toolbar is now in a single "line" and the stats are shown in a
single place.

The same is done for the PipelineDetailsPage where the pipeline stats
(trigger, management and result events) were previously listed right
next to the pipeline description which also took up a lot of space.

Change-Id: Ic3e3cbd3054c2bca7c1dd386e4b59d97905697da