Incorporate tenants stats in Toolbar section
Currently, the queue length and the last reconfiguration timestamp are
always shown on top of the status page in different places which takes
up a lot of space. Additionally, this information is quite important,
but oftentimes not relevant for most of the users.

So, instead of allocating a lot of space at the top of the page we could
utilize this space to show more of the "main" content and incorporate
the stats in the toolbar.

This change adds a new section to the toolbar that can be utilized to
show different kind of stats. Each stat is shown in form of a label with
a short summary (e.g. the number of events) and opens a Tooltip with
more detailed information when hovering over the label.

The design idea behind this change is that the toolbar is now split up
in two parts: A filter section on the left side and a button/stats group
on the right side. The filter section controls which elements are shown
on the status page whereby the button group provides additional
functionality like refreshing the page or showing tenant-specific stats.

Since I didn't find a way to split up a PF4 Toolbar component itself,
the "toolbar section" now contains two Toolbars which are split by a
Level layout. This ensures that the button/stats group stays on the
right side of the viewport and wraps to the next line if the viewport is
to small.

This makes the top of the status page much cleaner and more compact as
the toolbar is now in a single "line" and the stats are shown in a
single place.

The same is done for the PipelineDetailsPage where the pipeline stats
(trigger, management and result events) were previously listed right
next to the pipeline description which also took up a lot of space.

Change-Id: Ic3e3cbd3054c2bca7c1dd386e4b59d97905697da