Add "autohold_matches" var in job inventory
Add a variable called "autohold_matches". The variable is a list
of ids of autohold requests that could potentially trigger with the
current build.
This information can be used to take additional steps in a job should
a autohold trigger (prepare env, hold external resources not managed
by nodepool, etc). Note though that a autohold is not guaranteed to
trigger if the list isn't empty: the variable is evaluated at the
beginning of the build while zuul triggers a autohold at the very end
of it. This means that potentially, autohold requests might have expired
or be consumed in the time it took to run the build's play.
If users have enough control when managing autohold requests and zuul queue
events, this becomes a reliable indicator.

Change-Id: I87d807e7909ccb1559e38113bfd4c0a6bd1700ac