Add image/upload delete lifecycle
This adds support for deleting images and uploads.  This could be
commanded in the future through the API, or automatically if there
are more than 2 valid images.

It emits a delete-image trigger event so that a job can be run to
clean up artifact storage, but complete implementation and testing
of that is left for a later change.

Change-Id: Iea6f3c9f1f235576540717a4fcfbc0243038bd26