Use FilterIcon for filter dropdown on QueueItem
This replaces the SearchPlus icon that is currently used to open the
filter dropdown on a QueueItem card with a Filter icon.

A lot of users are confused by the SearchPlus icon. Since the icon is
used to apply different filters to the status page, using a Filter icon
makes more sense.

Change-Id: I7467e40438f1d42dca2c4c20d0fe87bef84380b5