Add labels and flavors to web
This adds an API endpoint for flavors, and updates the existing
API endpoint for labels to include niz-style labels.

If a client needs to differentiate, it can do so by detecting
the presesnce of a canonical_name field.  The web UI does not
need to differentiate; both types of labels will appear in the
existing labels tab.

A new tab is added to show flavors in the tenant.

The provider page is updated with sub-tabs for labels and flavors
relevant to the selected provider.

Change-Id: I40dd56d85c6f208929a7ce8339a9be2484df404d