Add niz nodes to rest api nodes list endpoint
This adds the new-style nodes (if any for the tenant) to the list
of nodes returned at the /nodes endpoint.  For the moment, it includes
mostly the same information as the existing nodes, plus a couple of
new fields that we may transition to in the future.  After the old
nodes are deprecated we can make the node listing display more
rich.

Change-Id: I02e5970917acc65138cd8ddd58e4395b57ca8e5e