Add information to start of job log
When a job starts, if there is a large number or size of repos,
then a user may be looking at the "Job console starting..." message
for a while without indication of progress.  Add a few more log
entries there so that users can see the progress through the
different phases.  We don't log each individual repo since that
may be excessively verbose (and even start to impact performance
at large scale).  We don't log individual repos in the Ansible
workspace prep roles either for the same reason.

Also, remove the "..." from the two existing log messages since
we now expect the next to appear sooner, and are not including them
on the new log messages sincethey are not necessary and introduce
noise and ambiguity to the log.

Change-Id: I4e2ef399ebe8ae62a31e237ceadf3f75de93f35c