Allow boolean claim rules
This change allows matching boolean claims in a claim rule. In order to
enable this, templating is now only applied to string values.

A case where this is useful is when matching accounts from an entire
Google Workspace domain. This is because Google states that we should
only trust tokens which have the email_verified claim set to true.[1]

[1]: https://developers.google.com/identity/gsi/web/guides/verify-google-id-token

Change-Id: Iaeec233644a058ebbdc510666c4b3e591cb0079e