Configure notify settings when reporting to gerrit
Since not all pipelines need to notify ALL (the default [1]) - allow the
pipeline config to specify the notification level via success/failure
action config.

* Rationale

The reason to do this is to avoid spamming everyone that gets
notifications for a change. For instance in the case where a vote is
not negative and the pipeline has mostly trivial jobs it would be nice
to avoid sending emails.

* Problems/also considered

This is potentially a breaking change if the votes that users have
configured are named "notify".

I would think the ultimate fix would be to, instead of having ad-hoc
parsers for votes, make the gerrit key have a ReviewInfo object (as
Yaml) as its value.

I don't want to do that as part of this review since that would be a
breaking change.

[1] https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#review-input

Change-Id: Ibf4d3462440f8aa10e78e7b773e04cc8d4497340