[WIP] sqlreporter: ensure build end data is stored even if log URL is too long
When a build's log URL is over 255 characters, replace it with a warning
message so that the build's end data can be stored - and retrieved
regardless.

The previous behavior would drop the data entry, with the drop event
being logged only if sqlalchemy's log level was appropriately set. This
left end users (and potentially zuul admins) confused as to what happened
to their builds - as they appeared in the web UI as in error state with
no end date regardless of their actual outcome.

The new behavior is to "gracefully" handle the error. This way, correct
data about the build's end is still stored in the database, and the
user is left to compute the log's URL by themselves (this is relatively
straightforward).

Change-Id: I3743bdd3415231ebcbe1b7ab20c5414c98273dd0