Fix dict mutation in web info endpoint
The info and tenant-info web endpoints return a copy of the
system-wide auth capabilities dictionary and mutate it to overlay
tenant tenant-specific overrides (such as a default realm or
read-only access).  The auth capabilities dict may have nested
dictionaries, and since the auth capabilities dict was only
shallow-copied, if any changes were made to the nested values,
they may be shared across tenants.  This may cause the info
endpoints to produce differing values depending on the order
of access.

Fortunately, this does not represent a vulnerability as these
values are not consulted internally to determine access, instead
they are for consumption by the web client in order to direct
behavior.  Incorrect values may cause confusing behavior (attempts
to access endpoints which then denied access, or invalid logins)
but would not allow unintended access.

Correct this and adjust a unit test to verify the behavior.

Change-Id: I9fd30a8990066bdf7425f13eb15c28ca3e70b7ed