Update patch set 11
Patch Set 11:

As far as merging, we strive to get reviews from more than one maintainer, so it may just need more time or a bit of lobbying in Matrix to bring it to the attention of more reviewers.

Where Zuul's use of Gerrit APIs is concerned, I believe it should be relying on the REST API here, as long as it's properly configured with the necessary credentials, but I might be wrong about that.

Patch-set: 11
Attention: {"person_ident":"Gerrit User 37624 \u003c37624@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}