Make job-list on status page more compact
This makes the job-list shown for each QueueItem more compact by
reducing the space between the different list elements. The change
applies to the Status and PipelineDetails page as both use the same
React component.

In addition to that, make the job result labels and the job progress
bars a little smaller so they fit better into the straiter job rows.

Change-Id: Id2971653527a77026c90c5a62494df79f7e9e04f