Act on image build artifact unlock instead of image uploads
This is the sequence at the end of the image build job:

1. Lock ImageBuildArtifact
2. Create ImageUploads for each endpoint
3. Unlock ImageBuildArtifact

If we start processing image uploads between step 2 and 3, we may
be unable to lock the IBA and therefore skip handling the upload.
Or another launcher may be handling a subset of the uploads, but
not all of them.

To address the first issue, we should only start processing when
we unlock the image build artifact (step 3).

To address the second issue, we will re-set the event
that triggers checking uplods if we fail to lock the IBA.

Change-Id: I609f2087ad6e64b89d991719d9fe605a00916578