Update patch set 11
Patch Set 11:

BTW, on the subject Jeremy raised about namespace collisions -- this schema was originally designed with full awareness of that risk and it was deemed small enough to go with the more user-friendly flat hierarchy we have.  I don't want to change it unless we start seeing a whole lot more options causing problems.

Patch-set: 11