Summarize named queues in default (collapsed) view on status page
This change aligns the visualization of named and unnamed queues in the
default (collapsed) view: All queues for a single pipeline will be
visualized as a single queue. This avoids that e.g. the gate pipeline
breaks the gallery layout in the collapsed view because the pipeline
card is higher than the others. In the expanded view, named queues are
still visualized independently.

For easier access to specific queues, they are now expanded in two
steps: First, expand the pipeline to "list" all named queues (basically
the previous default view). Second, expand individual queues.

The functionality of the "Expand all" switch is updated accordingly.

Change-Id: I7cc7c49d479b2da842599d6cc72d66125d1d982e