web: Add an option to sort pipelines
This adds a dropdown menu giving users an option to sort pipelines on
the PipelineOverview page, either by the initial implicit sorting,
alphabetically by name, or by the number of items in the pipelines.

This is to give users more control over how pipelines are displayed,
especially in large tenants with many pipelines.

Change-Id: I5475a1de77851351b2e776e78cffef5815d0c2f2