Don't invalidate zk context on stop event
Instead of aborting all zkobject operations when a component is stopped,
we should try to complete the current op and shutdown gracefully. The
stop event will only be used to determine if we should continue retrying
an operation or not.

Treating the stop event like a invalid ZK session can lead to data loss.

Change-Id: I16aa5a184934d5e3f451502e577301ad1d063feb