Fix issue with job forcing a dependency to run
This fixes an edge case not covered in
I7df2262881cb952b46b8a99ad5ccbe912c2e070b.

A job that is NOT executed can force a dependency to run even if that
dependency has file filters that don't match.  This change ensures
that the forced-run only happens if the dependent job matches the
change and will be run.

Change-Id: I26313415b95b247bc4e8c2314d92d5958b23e285