Force hard transitive dependencies to run
This is a further fix to I7df2262881cb952b46b8a99ad5ccbe912c2e070b
in addition to I26313415b95b247bc4e8c2314d92d5958b23e285.

This ensures that if job A matches and will run, and it has a
hard dependency on job B, which has a hard dependency on job C,
that all three jobs run.  Tests are added to cover this case for
both job updates and simple file matches on the terminal job.

The original change intended to do this but only functioned for
a single level, not transitive dependencies.

Neither of these should force soft dependencies to run, so test
coverage for those is added as well.

Change-Id: I3e696ec98da6dc0085e9e64faf8f80b26e57466d