Fix whitelabel authz
The web app authz logic did not use the tenant.apiPrefix when
fetching authz info after obtaining a token.  Instead, it constructed
the url on its own assuming a non-whitelabel configuration.  In
a typical whitelabel config, the URL would be duplicated
(/api/tenant/foo/tenant/foo/authorizations).

To correct this, use the tenant apiPrefix.  That should always be
available since it comes from the /info endpoint which provides
the basic authn information.

Change-Id: Ia6d8bc9b8e2d01df6f2edce45f0ec3c388ebb5b3