web: Fix null reference exception for cancelled builds
When trying to render the Manifest component of cancelled builds that
don't have a log_url but a zuul manifest as an artifact, we run into
null reference exceptions because the component tries to render links to
files from the manifest using the non-existing log_url.

A more elaborate fix would probably be to always report a log_url if we
have one (like in this case). So this is rather a quick workaround to
make the js frontend more robust.

Change-Id: I4c666245629c4b35d8ee7bef78993d523548ad57