Don't cancel completed build requests
With fail-fast enabled the following sequence of events can happen:

* two (here A and B) or more builds for the same buildset finish roughly
  at the same time and send a build-completed event
* build A failed and is the first event processed, which will lead to
  build B being cancelled
* since the build request for B is no longer locked (build already
  finished) the executor client will create a synthetic build-completed
  event
* the original build-completed event for build B is processed, including
  reporting of build result data and artifacts to the database
* item is dequeued and reported
* synthetic build-completed event for B is processed leading to the
  creation of a fake build that is then re-reported to the database,
  overwriting existing build information (e.g. end-time, log URL, ...)

Overwriting valid build information can lead to the build UI just
showing an error. The reason is that we might have existing artifacts
for a build in the DB. When there are artifacts the JS code expects the
log URL to be valid. Howerver, due to the re-reporting when processing a
synthetic build-completed event the log URL is set to null.

Co-Authored-By: Felix Edel <felix.edel@bmw.de>
Change-Id: I06dcf837d8af801c9168160be0e841d7fa2b6381