Ignore config hash when comparing frozen jobs
The `job.config_hash` also considers the `job.branches` attribute.
However, when the job is frozen we only need to compare execution
related attributes when deciding whether a job can be deduplicated.

This means we can exclude the config hash when comparing jobs as we
already do for e.g. the inheritance path.

Change-Id: I3e51024812aba41c7f51db121fb4a5d426ee5938