UI: Include queue.branch in expansionKey for queues on status page
This fixes a bug in the queue expansion for queues containing a branch
name: If multiple queues with the same name but different branches are
visible within the same pipeline on the status page, all of them will be
expanded as they all share the same expansionKey. This can be fixed by
including the branch name in the expansionKey, so that only one specific
queue matches the key.

Change-Id: I5a5c9bf7faea485ffda92122cf3f3ae775b9774a