web: Scope pipeline/queue expansion by tenant and queue branch
When expanding a queue that exists for multiple branches, all those
queues are expanded, not just the one being clicked on. This is because
we store the expansion state only based on the queue name.

Fix this by including the branch name in the state key. The branch name
might be empty, in which case the last part of the state key is also
empty.

This change also incorporates the tenant name in the state key of queue
and pipeline expansions. This is to emphasise that pipelines are tenant
scoped.

Change-Id: I7243118083dab8344d4e64426b083fd0d503d5cc