Add debug log when fetching Gerrit branches
We've seen an instance where after a branch creation, the new
branch did not appear in the project.  This should help confirm
exactly when branches are fetched, and by logging the number of
branches, we can confirm the inclusion of the new branch.

We don't log the contents in order to avoid logspam from projects
with thousands of branches.

Change-Id: Ifb296acb7292e2274cfc7729d87139440cf7c325