Fix bulk repo state restore
It looks like I2727b00d540ed0d6d650342f427503ddcf5cd533 introduced a
regression in that it attempts to restore the full repo state of all
projects in the buildset, and not only the projects relevant to
the specific job.

Since the job only clones/updates projects that are relevant
(part of `args['project']`) this can lead to failures e.g. when the
repo doesn't exist on the executor. In other cases we might just do
clones/checkouts that are not necessary.

Additionally, projects that are used by include-vars were not included
in the `args['projects']` list, meaning that with this fix in place,
they might not be prepared properly.  This change also corrects that.

Change-Id: Ia4a297fb1c30271a9babb25acff4727faf464aa1