web: Give root element whole viewport height
On the status page, some overlay elements like the filter dropdown can
appear cropped because the height of the apps root element is based on
the height of the child container. With only a few pipelines being
displayed, this might not be enough to display the entire dropdown list.

This change fixes this by setting the height of the root element to 100%
of the viewport explicitly.

Change-Id: Ie07975d67de4c23adf09c7eab9471599ec97a3a3