Retry nodescan connections during the key phase
The nodescan state machines has 4 main states:
* initial connection
* initial key negotiation
* key connection
* key negotiation

We expect many connection failures during the "initial connection"
phase, but if we encounter a hung connection during the
"key connection" phase, then we will wait for the complete boot
timeout duration and then fail.  In an attempt to reduce nodescan
failures due to initial flaky connections, retry the connection
attempt for the same key if we encounter a failure during connection
(but not negotiation -- a negotiation failure will still proceed to
the next key).  In addition, allocate 10 seconds for each key
connection attempt before timing out and retrying the connection.
This matches an existing 10 second connection timeout we have in the
initial connection attempt.

The _connect() method is updated so that it never sets the next state,
instead, we will set that at the call sites.  It is important that
once we pass the negotiating_init state that we don't go back to
it, since we may have started to restrict the set of keys.  Setting
the next state at the call sites makes that explicit.

Change-Id: I34691af325ce806e2bf0ae658da09aed0239d82d